Switch from pyes to Elasticsearch 7
pyes doesn't seem to be supported anymore, and all the functionality we
want is provided by the Elasticsearch client we already have installed.

This also involves some updates to be compatible with Elasticsearch 7,
which is the latest open source version. The main thing is the "string"
type is now called "text".