Ensure instance_path is relative to root_path
When running locally, the app was trying to load things from /instance
(as in, the root of the filesystem). This isn't what we want, so
prepending app.root_path to fix it.

I'm not sure _which_ upgrade broke this functionality, but the fix seems
like it will be robust enough.